+++ src/Swarm/Game/Recipe.hs
@@ -49,8 +49,8 @@ import Data.Text (Text)
 import Data.Text qualified as T
 import Data.Yaml
 import GHC.Generics (Generic)
-import Swarm.Failure
 import Swarm.Game.Entity as E
+import Swarm.Game.Failure
 import Swarm.Game.ResourceLoading (getDataFileNameSafe)
 import Swarm.Util.Yaml
 import Witch
