+++ src/Swarm/Game/ResourceLoading.hs
@@ -14,7 +14,7 @@ import Data.Maybe (mapMaybe)
 import Data.Text (Text)
 import Data.Text qualified as T
 import Paths_swarm (getDataDir)
-import Swarm.Failure
+import Swarm.Game.Failure
 import Swarm.Util
 import System.Directory (
   XdgDirectory (XdgData),
