+++ src/Swarm/TUI/Model/UI.hs
@@ -58,11 +58,12 @@ import Data.Map (Map)
 import Data.Map qualified as M
 import Data.Text (Text)
 import Data.Text qualified as T
-import Swarm.Failure (SystemFailure)
-import Swarm.Failure.Render (prettyFailure)
 import Swarm.Game.Achievement.Attainment
 import Swarm.Game.Achievement.Definitions
 import Swarm.Game.Achievement.Persistence
+import Swarm.Game.Failure (SystemFailure)
+import Swarm.Game.Failure.Render (prettyFailure)
+import Swarm.Game.ResourceLoading (getSwarmHistoryPath, readAppData)
 import Swarm.Game.ScenarioInfo (
   ScenarioInfoPair,
  )
@@ -74,7 +75,6 @@ import Swarm.TUI.Model.Menu
 import Swarm.TUI.Model.Name
 import Swarm.TUI.Model.Repl
 import Swarm.Util
-import Swarm.Game.ResourceLoading (getSwarmHistoryPath, readAppData)
 import System.Clock
 
 ------------------------------------------------------------
